# specs/017-google-authentication/contracts/auth-api-contract.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Simo Authentication API",
    "version": "1.0.0",
    "description": "Google OAuth and Session Management Endpoints for Simo Ecosystem"
  },
  "paths": {
    "/api/v1/auth/google": {
      "post": {
        "summary": "Exchange Google ID Token for Simo Session JWT",
        "operationId": "authenticateWithGoogle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id_token"],
                "properties": {
                  "id_token": {
                    "type": "string",
                    "description": "Google ID Token returned from Google Sign-In SDK"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string" },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "email": { "type": "string" },
                            "display_name": { "type": "string" },
                            "avatar_url": { "type": "string", "nullable": true }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired Google Token"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Get authenticated user profile",
        "operationId": "getAuthenticatedUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "display_name": { "type": "string" },
                        "avatar_url": { "type": "string", "nullable": true },
                        "created_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}